requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
idna==3.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.11
